fixed some warnings in ECCSF

+++ app/src/main/java/com/example/finalproject/ECCSFfav.java
@@ -3,7 +3,6 @@ package com.example.finalproject;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.content.ContentValues;
 import android.content.Intent;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
@@ -20,11 +19,11 @@ import android.widget.Toast;
 import java.util.ArrayList;
 
 public class ECCSFfav extends AppCompatActivity {
-    private ArrayList<ChargingStation> favStations = new ArrayList<ChargingStation>();
+    private ArrayList<ChargingStation> favStations = new ArrayList<>();
     private static ECCSFDatabaseOpenHelper dbOpener;
     private static SQLiteDatabase db;
     private static BaseAdapter myAdapter;
-    private Intent delData= new Intent();;
+    private Intent delData= new Intent();
     private int numOfDeleted;
 
     @Override
@@ -100,51 +99,54 @@ public class ECCSFfav extends AppCompatActivity {
     @Override
     protected void onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
+        if(data!=null){
+            // from detail page, to delete one station from fav
+            if (requestCode == 5 && resultCode == 3) {
+                // if the data is needed to be deleted from favorite stations
+                if (data.getBooleanExtra("deleteFromFav", false)) {
+                    String latitude = data.getStringExtra("latitude");
+                    String longitude = data.getStringExtra("longitude");
 
-        // from detail page, to delete one station from fav
-        if (requestCode == 5 && resultCode == 3) {
-            // if the data is needed to be deleted from favorite stations
-            if (data.getBooleanExtra("deleteFromFav", false)) {
-                String latitude = data.getStringExtra("latitude");
-                String longitude = data.getStringExtra("longitude");
+                    Cursor cursor = db.query(true, ECCSFDatabaseOpenHelper.TABLE_NAME,
+                            new String[]{ECCSFDatabaseOpenHelper.COL_ID},
+                            ECCSFDatabaseOpenHelper.COL_LATITUDE + " = ? AND " +
+                                    ECCSFDatabaseOpenHelper.COL_LONGITUDE + " = ? "
+                            , new String[]{latitude, longitude}, null, null, null, null);
+                    cursor.moveToFirst();
+                    if (cursor.getCount() > 0) {
+                        int id = cursor.getInt(cursor.getColumnIndex(ECCSFDatabaseOpenHelper.COL_ID));
+                        Log.i("id to be deleted is" + id, "eccsfmain");
+                        db.delete(ECCSFDatabaseOpenHelper.TABLE_NAME, ECCSFDatabaseOpenHelper.COL_ID + "=?",
+                                new String[]{Long.toString(id)});
+                    }
+                    cursor.close();
 
-                Cursor cursor = db.query(true, ECCSFDatabaseOpenHelper.TABLE_NAME,
-                        new String[]{ECCSFDatabaseOpenHelper.COL_ID},
-                        ECCSFDatabaseOpenHelper.COL_LATITUDE + " = ? AND " +
-                                ECCSFDatabaseOpenHelper.COL_LONGITUDE + " = ? "
-                        , new String[]{latitude, longitude}, null, null, null, null);
-                cursor.moveToFirst();
-                if (cursor.getCount() > 0) {
-                    int id = cursor.getInt(cursor.getColumnIndex(ECCSFDatabaseOpenHelper.COL_ID));
-                    Log.i("id to be deleted is" + id, "eccsfmain");
-                    db.delete(ECCSFDatabaseOpenHelper.TABLE_NAME, ECCSFDatabaseOpenHelper.COL_ID + "=?",
-                            new String[]{Long.toString(id)});
-                }
-                cursor.close();
+                    for(ChargingStation station: favStations){
+                        if(station.getLongitude().equals(longitude)){
+                            favStations.remove(station);
+                            Toast.makeText(this, "Successfully deleted from favorite stations",
+                                    Toast.LENGTH_LONG).show();
+                            myAdapter.notifyDataSetChanged();
 
-                for(ChargingStation station: favStations){
-                    if(station.getLongitude().equals(longitude)){
-                        favStations.remove(station);
-                        Toast.makeText(this, "Successfully deleted from favorite stations",
-                                Toast.LENGTH_LONG).show();
-                        myAdapter.notifyDataSetChanged();
+                            delData.putExtra("numOfDel", numOfDeleted);
+                            delData.putExtra(numOfDeleted+"", latitude);
+                            numOfDeleted++;
+                            break;
+                        }
+                    }
 
-                        delData.putExtra("numOfDel", numOfDeleted);
-                        delData.putExtra(numOfDeleted+"", latitude);
-                        numOfDeleted++;
-                        break;
+
+                    View emptyView = findViewById(R.id.empty_list);
+                    if(myAdapter.getCount() == 0){
+
+                        emptyView.setVisibility(View.VISIBLE);
+                    }else{
+                        emptyView.setVisibility(View.GONE);
                     }
                 }
+        }
 
 
-                View emptyView = findViewById(R.id.empty_list);
-                if(myAdapter.getCount() == 0){
-
-                    emptyView.setVisibility(View.VISIBLE);
-                }else{
-                    emptyView.setVisibility(View.GONE);
-                }
-            }
         }
     }
     class MyListAdapter extends BaseAdapter {
@@ -165,17 +167,17 @@ public class ECCSFfav extends AppCompatActivity {
 
         @Override
         public View getView(int position, View oldView, ViewGroup parent) {
-            View thisView = oldView;
-            thisView = getLayoutInflater().inflate(R.layout.eccsf_listview_row, parent, false);
 
-            TextView title = thisView.findViewById(R.id.row_title);
-            TextView address = thisView.findViewById(R.id.row_address);
-            TextView saved = thisView.findViewById(R.id.row_saved);
+            oldView = getLayoutInflater().inflate(R.layout.eccsf_listview_row, parent, false);
+
+            TextView title = oldView.findViewById(R.id.row_title);
+            TextView address = oldView.findViewById(R.id.row_address);
+            TextView saved = oldView.findViewById(R.id.row_saved);
             saved.setVisibility(View.INVISIBLE);
 
             title.setText(favStations.get(position).getTitle());
             address.setText(favStations.get(position).getAddress());
-            return thisView;
+            return oldView;
         }
     }
 }
